Embedding/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-genai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
transformers>=4.30.0
torch>=2.0.0
